pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coldcast"
version = "0.1.0"
description = "Weather and hydrologic data downloads"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Coldcast Contributors" }]
dependencies = [
  "pandas>=2.0",
  "pyyaml>=6.0",
  "requests>=2.31",
  "cdsapi>=0.7",
  "ecmwf-opendata>=0.3",
  "netCDF4>=1.6",
  "xarray>=2023.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[project.scripts]
coldcast = "coldcast.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
coldcast = ["data/*.yaml", "data/*.csv"]

[tool.pytest.ini_options]
markers = [
  "network: tests that require network access",
  "api_client: api client integration tests (mocked clients)",
]
